Refactor post widget and fix sliverappbar layout
